+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add `--output` option to optionally specify output file
+- Add `--format` option to optionally specify output format (LLHD or MLIR)
+
 ### Changed
 - Update VHDL crate to same `llhd` version as SystemVerilog